Oops... Long is limited...

+++ src-ui/sdr/airspy.h
@@ -22,7 +22,7 @@ public:
     void stop();
     void drawUI();
     static void init();
-    virtual void setFrequency(long frequency);
+    virtual void setFrequency(float frequency);
     static std::vector<std::tuple<std::string, sdr_device_type, uint64_t>> getDevices();
 };
 #endif
